Add-on crash fixed.
